+++ pyproject.toml
@@ -25,34 +25,32 @@ requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-dv-banner = "datastation.scripts.dv_banner:main"
-dv-dataset-oai-harvest = "datastation.scripts.oai_harvest:main"
-dv-datacite-records-update = "datastation.scripts.update_datacite_records:main"
-dv-dataset-add-role-assignment="datastation.scripts.add_role_assignments:main"
-dv-dataset-delete-draft="datastation.scripts.delete_draft_datasets:main"
-dv-dataset-delete-role-assignment="datastation.scripts.delete_role_assignments:main"
-dv-dataset-destroy="datastation.scripts.dataset_destroy:main"
-dv-dataset-destroy-migration-placeholder="datastation.scripts.dataset_destroy_migration_placeholder:main"
-dv-dataset-find-with-role-assignment="datastation.scripts.find_datasets_with_roleassignment:main"
-dv-dataset-publish="datastation.scripts.publish_datasets:main"
-dv-dataset-reindex="datastation.scripts.reindex_datasets:main"
-dv-dataset-replace-metadata-field-values = "datastation.scripts.replace_metadata_field_values:main"
-dv-dataset-retrieve-metadata="datastation.scripts.retrieve_dataset_metadata:main"
-dv-dataset-retrieve-metadata-field="datastation.scripts.retrieve_dataset_metadata_field:main"
-dv-dataset-locks = "datastation.scripts.manage_dataset_lock:main"
-dv-dataset-verify = "datastation.scripts.verify_dataset:main"
-dv-dataverse-oai-harvest = "datastation.scripts.oai_harvest:main"
-dv-dataverse-retrieve-pids="datastation.scripts.retrieve_dataset_pids:main"
-dv-file-prestage="datastation.scripts.prestage_files:main"
-dv-user-import="datastation.scripts.import_user:main"
-
-ingest-flow-copy-batch-to-ingest-area="datastation.scripts.ingest_flow_copy_batch_to_ingest_area:main"
-ingest-flow-move-batch-to-ingest-area="datastation.scripts.ingest_flow_move_batch_to_ingest_area:main"
-ingest-flow-start-migration="datastation.scripts.ingest_flow_start_migration:main"
-ingest-flow-start-import="datastation.scripts.ingest_flow_start_import:main"
-ingest-flow-list-events="datastation.scripts.ingest_flow_list_events:main"
-ingest-flow-progress-report="datastation.scripts.ingest_flow_progress_report:main"
-ingest-flow-block="datastation.scripts.ingest_flow_block:main"
-ingest-flow-unblock="datastation.scripts.ingest_flow_unblock:main"
-
-dans-bag-validate="datastation.scripts.validate_dans_bag:main"
+dans-bag-validate="datastation.old.scripts.validate_dans_bag:main"
+dv-banner = "datastation.old.scripts.dv_banner:main"
+dv-datacite-records-update = "datastation.old.scripts.update_datacite_records:main"
+dv-dataset-add-role-assignment="datastation.old.scripts.add_role_assignments:main"
+dv-dataset-delete-draft="datastation.old.scripts.delete_draft_datasets:main"
+dv-dataset-delete-role-assignment="datastation.old.scripts.delete_role_assignments:main"
+dv-dataset-destroy-migration-placeholder="datastation.old.scripts.dataset_destroy_migration_placeholder:main"
+dv-dataset-destroy="datastation.old.scripts.dataset_destroy:main"
+dv-dataset-find-with-role-assignment="datastation.old.scripts.find_datasets_with_roleassignment:main"
+dv-dataset-locks = "datastation.old.scripts.manage_dataset_lock:main"
+dv-dataset-oai-harvest = "datastation.old.scripts.oai_harvest:main"
+dv-dataset-publish="datastation.old.scripts.publish_datasets:main"
+dv-dataset-reindex="datastation.old.scripts.reindex_datasets:main"
+dv-dataset-replace-metadata-field-values = "datastation.old.scripts.replace_metadata_field_values:main"
+dv-dataset-retrieve-metadata-field="datastation.old.scripts.retrieve_dataset_metadata_field:main"
+dv-dataset-retrieve-metadata="datastation.old.scripts.retrieve_dataset_metadata:main"
+dv-dataset-verify = "datastation.old.scripts.verify_dataset:main"
+dv-dataverse-oai-harvest = "datastation.old.scripts.oai_harvest:main"
+dv-dataverse-retrieve-pids="datastation.old.scripts.retrieve_dataset_pids:main"
+dv-file-prestage="datastation.old.scripts.prestage_files:main"
+dv-user-import="datastation.old.scripts.import_user:main"
+ingest-flow-block="datastation.old.scripts.ingest_flow_block:main"
+ingest-flow-copy-batch-to-ingest-area="datastation.old.scripts.ingest_flow_copy_batch_to_ingest_area:main"
+ingest-flow-list-events="datastation.old.scripts.ingest_flow_list_events:main"
+ingest-flow-move-batch-to-ingest-area="datastation.old.scripts.ingest_flow_move_batch_to_ingest_area:main"
+ingest-flow-progress-report="datastation.old.scripts.ingest_flow_progress_report:main"
+ingest-flow-start-import="datastation.old.scripts.ingest_flow_start_import:main"
+ingest-flow-start-migration="datastation.old.scripts.ingest_flow_start_migration:main"
+ingest-flow-unblock="datastation.old.scripts.ingest_flow_unblock:main"
