+++ src/main/datastation/old/scripts/dataset_destroy.py
@@ -1,8 +1,8 @@
 import argparse
 import logging
 
-from datastation.config import init
-from datastation.dv_api import destroy_dataset
+from datastation.old.config import init
+from datastation.old.dv_api import destroy_dataset
 
 
 def main():
