+++ src/main/datastation/old/scripts/delete_draft_datasets.py
@@ -2,10 +2,10 @@ import sys
 import argparse
 import logging
 
-from datastation.batch_processing import batch_process
-from datastation.config import init
-from datastation.ds_pidsfile import load_pids
-from datastation.dv_api import delete_dataset_draft
+from datastation.old.batch_processing import batch_process
+from datastation.old.config import init
+from datastation.old.ds_pidsfile import load_pids
+from datastation.old.dv_api import delete_dataset_draft
 
 
 def delete_dataset_command(server_url, api_token, pids_file, delay):
