+++ src/main/datastation/old/scripts/ingest_flow_list_events.py
@@ -1,6 +1,6 @@
 import argparse
-from datastation.ingest_flow import start_import, list_events
-from datastation.config import init
+from datastation.old.ingest_flow import list_events
+from datastation.old.config import init
 
 
 def main():
