+++ src/main/datastation/old/scripts/reindex_datasets.py
@@ -1,9 +1,9 @@
 import argparse
 
-from datastation.batch_processing import batch_process
-from datastation.config import init
-from datastation.ds_pidsfile import load_pids
-from datastation.dv_api import reindex_dataset
+from datastation.old.batch_processing import batch_process
+from datastation.old.config import init
+from datastation.old.ds_pidsfile import load_pids
+from datastation.old.dv_api import reindex_dataset
 
 
 def reindex_dataset_command(server_url, delay, pids_file):
