+++ src/main/datastation/old/scripts/replace_metadata_field_values.py
@@ -3,11 +3,11 @@ import logging
 import json
 import sys
 
-from datastation.batch_processing import batch_process
-from datastation.config import init
-from datastation.ds_pidsfile import load_pids
+from datastation.old.batch_processing import batch_process
+from datastation.old.config import init
+from datastation.old.ds_pidsfile import load_pids
 
-from datastation.dv_api import replace_dataset_metadatafield, get_dataset_metadata
+from datastation.old.dv_api import replace_dataset_metadatafield, get_dataset_metadata
 
 
 def replace_metadatafield(server_url, api_token, pid, updated_field):
